requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
web3==6.15.1
google-generativeai==0.5.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.4
